requirements.txt
Django==2.1.7
djangorestframework==3.9.1
requests==2.21.0
django-environ==0.4.5
huey==1.11.0
bencode.py==2.0.0
psycopg2-binary==2.7.7
Markdown==3.0.1
redis==3.2.0
coreapi==2.3.3
